+++ samples/TestSystemAlbum/src/main/java/com/leachchen/testsystemalbum/Bean/PhoneVideoBean.java
@@ -40,7 +40,7 @@ public class PhoneVideoBean {
      }
      */
 
-    private int _id;
+    private long _id;
     private String _data;
     private String _display_name;
     private long _size;
@@ -70,7 +70,7 @@ public class PhoneVideoBean {
 
     public PhoneVideoBean(Cursor cursor)
     {
-        this._id= cursor.getInt(0);
+        this._id= cursor.getLong(0);
         this._data= cursor.getString(1);
         this._display_name= cursor.getString(2);
         this._size= cursor.getLong(3);
@@ -99,11 +99,11 @@ public class PhoneVideoBean {
         this.height= cursor.getInt(26);
     }
 
-    public int get_id() {
+    public long get_id() {
         return _id;
     }
 
-    public void set_id(int _id) {
+    public void set_id(long _id) {
         this._id = _id;
     }
